add modules subtask
